# spec/openapi.yaml
openapi: 3.1.0
info:
  title: arw-server
  description: Unified ARW server API surface (headless-first).
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  version: 0.1.4
paths:
...
  /admin/memory/quarantine:
    get:
      tags:
      - Review
      operationId: memory_quarantine_get_doc
      responses:
        '200':
          description: Memory quarantine entries
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Review: GET /admin/memory/quarantine.'
    post:
      tags:
      - Review
      operationId: memory_quarantine_queue_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryQuarantineRequest'
        required: true
      responses:
        '200':
          description: Queued for review
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
        '500':
          description: Storage error
      description: 'Review: POST /admin/memory/quarantine.'
  /admin/memory/quarantine/admit:
    post:
      tags:
      - Review
      operationId: memory_quarantine_admit_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryQuarantineAdmit'
        required: true
      responses:
        '200':
          description: Entry removed
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
        '500':
          description: Storage error
      description: 'Review: POST /admin/memory/quarantine/admit.'
...
  /admin/world_diffs:
    get:
      tags:
      - Review
      operationId: world_diffs_get_doc
      responses:
        '200':
          description: Queued world diffs
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Review: GET /admin/world_diffs.'
  /admin/world_diffs/decision:
    post:
      tags:
      - Review
      operationId: world_diffs_decision_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorldDiffDecision'
        required: true
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
        '404':
          description: Diff not found
        '500':
          description: Storage error
      description: 'Review: POST /admin/world_diffs/decision.'
  /admin/world_diffs/queue:
    post:
      tags:
      - Review
      operationId: world_diffs_queue_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorldDiffQueueRequest'
        required: true
      responses:
        '200':
          description: Diff queued
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
        '500':
          description: Storage error
      description: 'Review: POST /admin/world_diffs/queue.'
...
  /research_watcher/{id}/approve:
    post:
      tags:
      - Research
      operationId: research_watcher_approve_doc
      parameters:
      - name: id
        in: path
        description: Watcher item id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatcherDecision'
        required: true
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Error
      description: 'Research: POST /research_watcher/{id}/approve.'
  /research_watcher/{id}/archive:
    post:
      tags:
      - Research
      operationId: research_watcher_archive_doc
      parameters:
      - name: id
        in: path
        description: Watcher item id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatcherDecision'
        required: true
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Error
      description: 'Research: POST /research_watcher/{id}/archive.'
...
components:
  schemas:
...
    MemoryQuarantineAdmit:
      type: object
      required:
      - id
      properties:
        id:
          type: string
    MemoryQuarantineRequest:
      type: object
      properties:
        content_preview:
          type:
          - string
          - 'null'
        content_type:
          type:
          - string
          - 'null'
        evidence_score:
          type:
          - number
          - 'null'
          format: double
        extractor:
          type:
          - string
          - 'null'
        id:
          type:
          - string
          - 'null'
        project_id:
          type:
          - string
          - 'null'
        provenance:
          type:
          - string
          - 'null'
        risk_markers:
          type:
          - array
          - 'null'
          items:
            type: string
...
    WatcherDecision:
      type: object
      properties:
        note:
          type:
          - string
          - 'null'
...
    WorldDiffDecision:
      type: object
      required:
      - id
      - decision
      properties:
        decision:
          type: string
        id:
          type: string
        note:
          type:
          - string
          - 'null'
    WorldDiffQueueRequest:
      type: object
      properties:
        changes: {}
        from_node:
          type:
          - string
          - 'null'
        id:
          type:
          - string
          - 'null'
        project_id:
          type:
          - string
          - 'null'
        summary:
          type:
          - string
          - 'null'
...